+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -4,6 +4,7 @@ import static kr.codesqaud.cafe.common.consts.SessionConst.SESSION_USER;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +14,8 @@ import org.springframework.web.bind.annotation.SessionAttribute;
 
 import kr.codesqaud.cafe.controller.dto.req.JoinRequest;
 import kr.codesqaud.cafe.controller.dto.req.ProfileEditRequest;
+import kr.codesqaud.cafe.exception.DuplicatedUserIdException;
+import kr.codesqaud.cafe.exception.InvalidPasswordException;
 import kr.codesqaud.cafe.service.UserService;
 
 @Controller
@@ -58,5 +61,17 @@ public class UserController {
 		userService.validateHasAuthorization(sessionUserId, userId);
 		userService.editUserProfile(userId, request);
 		return "redirect:/users";
+	}
+
+	@ExceptionHandler(DuplicatedUserIdException.class)
+	public String handleDuplicatedUserId(final DuplicatedUserIdException e, final Model model) {
+		model.addAttribute("error", e.getMessage());
+		return "user/form";
+	}
+
+	@ExceptionHandler(InvalidPasswordException.class)
+	public String handleInvalidPassword(final InvalidPasswordException e, final Model model) {
+		model.addAttribute("error", e.getMessage());
+		return "user/edit_form";
 	}
 }
