+++ src/main/java/kr/codesqaud/cafe/exception/GlobalExceptionHandler.java
@@ -7,22 +7,10 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-	@ExceptionHandler(DuplicatedUserIdException.class)
-	public String handleDuplicatedUserId(final DuplicatedUserIdException e, final Model model) {
-		model.addAttribute("error", e.getMessage());
-		return "user/form";
-	}
-
 	@ExceptionHandler(NotFoundException.class)
 	public String handleUserNotFound(final NotFoundException e, final Model model) {
 		model.addAttribute("error", e.getMessage());
 		return "user/notfound";
-	}
-
-	@ExceptionHandler(InvalidPasswordException.class)
-	public String handleInvalidPassword(final InvalidPasswordException e, final Model model) {
-		model.addAttribute("error", e.getMessage());
-		return "user/edit_form";
 	}
 
 	@ExceptionHandler(InvalidLoginInfoException.class)
